+++ CHANGES.rst
@@ -5,6 +5,10 @@ IP Navigator CHANGES
 
 Development
 ===========
+
+
+2019-05-13 0.166.0
+==================
 - [ui] Upgrade to jQuery 3.4.1
 - [ui] Add "clean-webpack-plugin" to cleanup the assets folder before building
 - [mw] Improve settings for having per-vendor OPS credentials